+++ docs/develop.md
@@ -16,7 +16,7 @@ See [./dev-install-from-source.md](./dev-install-from-source.md)
 
 Run the tests with `npm run test`.
 
-There's also a markdown converter test you can run with `npm run md-diff` (requires [nodemon](https://www.npmjs.com/package/nodemon); `npm install -g nodemon`) which creates a file named `md.diff.html` that displays the difference between the markdown converter's actual output and its expected output. Any text with a green background is missing from the actual output, and any text with a red background is unexpected. Any incorrect newline character is visualized as a downward right arrow (⤵). You may want to use VS Code's [Live Server](https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer) extension to automatically reload the output HTML file when nodemon changes it. Note that VS Code sometimes makes automatic changes to markdown files while they're being edited that could mess up `expected.md`, such as making each table column have a consistent width across rows. If you edit `expected.md` in VS Code, please fix the changes VS Code makes before committing.
+If a certain test fails, its error message will tell you to run `npm run md-diff` (requires [nodemon](https://www.npmjs.com/package/nodemon); `npm install -g nodemon`) and open a file named `md.diff.html` that displays the differences between the markdown converter's actual output and its expected output. Any text with a green background is missing from the actual output, and any text with a red background is unexpected. Any incorrect newline character is visualized as a downward right arrow (⤵). You may want to use VS Code's [Live Server](https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer) extension to automatically reload `md.diff.html` when nodemon changes it. Note that VS Code sometimes makes automatic changes to markdown files while they're being edited that could mess up `expected.md`, such as changing table column widths or ordered list numbers. If you edit `expected.md` in VS Code, please fix any changes VS Code automatically makes before committing.
 
 ## Git workflow for collaboration
