+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "test": "node --test",
-    "md-diff": "nodemon ./test/md.js --ext * --ignore *.diff.html",
+    "md-diff": "nodemon ./test/diffMd.js --ext * --ignore *.diff.html",
     "dev-firefox": "rollup --config rollup.config.firefox.mjs",
     "build-firefox": "npm install && npm run dev-firefox && npm run assert-clean && rm -r src && rm -r chrome",
     "update-firefox": "git restore . && git pull && npm run build-firefox",
